(PC-18448)[PRO] refactor: remove filtered routes objects and rename files
