+++ pro/src/pages/Signup/Signup.tsx
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react'
 import { Routes, Route } from 'react-router-dom'
 
 import AppLayout from 'app/AppLayout'
-import { IRoute } from 'app/AppRouter/routes_map'
-import { routesSignup } from 'app/AppRouter/subroutes_signup_map'
+import { IRoute } from 'app/AppRouter/routesMap'
+import { routesSignup } from 'app/AppRouter/subroutesSignupMap'
 import PageTitle from 'components/PageTitle/PageTitle'
 import useActiveFeature from 'hooks/useActiveFeature'
 import { campaignTracker } from 'tracking/mediaCampaignsTracking'
