+++ pro/src/pages/SignupJourneyRoutes/SignupJourneyRoutes.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { Route, Routes, useLocation, useNavigate } from 'react-router-dom'
 
 import AppLayout from 'app/AppLayout'
-import { IRoute } from 'app/AppRouter/routes_map'
-import { routesSignupJourney } from 'app/AppRouter/subroutes_signupjourney_map'
+import { IRoute } from 'app/AppRouter/routesMap'
+import { routesSignupJourney } from 'app/AppRouter/subroutesSignupJourneyMap'
 import { SignupJourneyFormLayout } from 'components/SignupJourneyFormLayout'
 import { SignupJourneyContextProvider } from 'context/SignupJourneyContext'
 import { Events } from 'core/FirebaseEvents/constants'
